backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
numpy<2
soundcard
SpeechRecognition
openai-whisper
openai
python-dotenv
